requirements.txt
altair==5.4.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
easyocr==1.7.1
faiss-cpu==1.8.0.post1
filelock==3.16.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.24.6
idna==3.8
imageio==2.35.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.6.3
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pillow==10.4.0
protobuf==5.28.0
pyarrow==17.0.0
pyclipper==1.3.0.post5
pydeck==0.9.1
Pygments==2.18.0
PyPDF2==3.0.1
python-bidi==0.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
streamlit==1.38.0
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tokenizers==0.19.1
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.34.0
watchdog==4.0.2